Fixes, better beams.

+++ music/encoder.h
@@ -30,5 +30,7 @@ struct Encoder {
     /// Writes a video frame
     void writeVideoFrame(const Image& image);
     /// Writes an audio frame
-    void writeAudioFrame(ref<float2> audio);
+    //void writeAudioFrame(ref<float2> audio);
+    /// Writes an audio frame
+    void writeAudioFrame(ref<short2> audio);
 };
